# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EcoLearn Mauritius API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all published courses
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course with lessons and objectives
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments for the current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Enroll in a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    get:
      operationId: getEnrollment
      tags: [enrollments]
      summary: Get a single enrollment with progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentDetail"

  # ── PROGRESS ──────────────────────────────────────────────────────────────────
...
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  # ── TESTIMONIALS ──────────────────────────────────────────────────────────────
...
  /platform-admin/learning-paths/{id}/courses:
    put:
      operationId: platformAdminUpdateLearningPathCourses
      tags: [platform-admin-learning-paths]
      summary: Set path courses assignment and order positions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformAdminUpdateLearningPathCoursesInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  # ── PLATFORM ADMIN SDG GOALS & TARGETS ───────────────────────────────────────
...
components:
  schemas:
...
    Course:
      type: object
      required: [id, title, description, categoryId, durationMinutes, priceUsd, level, isFeatured, thumbnailUrl]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: ["string", "null"]
        description:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        priceUsd:
          type: number
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        isFeatured:
          type: boolean
        thumbnailUrl:
          type: ["string", "null"]
        previewVideoUrl:
          type: ["string", "null"]
        learningObjectives:
          type: array
          items:
            type: string
        enrollmentCount:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        includesCertificate:
          type: boolean
        passingScore:
          type: integer
        isPublished:
          type: boolean
        intendedRoles:
          type: array
          items:
            type: string
        version:
          type: integer
        reviewDate:
          type: ["string", "null"]
        recommendedNextCourseId:
          type: ["integer", "null"]
        createdAt:
          type: string
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/Lesson"
    CourseInput:
      type: object
      required: [title, description, categoryId, durationMinutes, priceUsd, level]
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        durationMinutes:
          type: integer
        priceUsd:
          type: number
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        isFeatured:
          type: boolean
        thumbnailUrl:
          type: ["string", "null"]
        previewVideoUrl:
          type: ["string", "null"]
        learningObjectives:
          type: array
          items:
            type: string
        includesCertificate:
          type: boolean
    CourseUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        durationMinutes:
          type: ["integer", "null"]
        priceUsd:
          type: ["number", "null"]
        level:
          type: ["string", "null"]
        isFeatured:
          type: ["boolean", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        previewVideoUrl:
          type: ["string", "null"]
        learningObjectives:
          type: array
          items:
            type: string
        includesCertificate:
          type: ["boolean", "null"]
    Lesson:
      type: object
      required: [id, courseId, title, orderIndex, durationMinutes]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        orderIndex:
          type: integer
        durationMinutes:
          type: integer
        videoUrl:
          type: ["string", "null"]
        pdfUrl:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        isArchived:
          type: boolean
        contentBlocks:
          type: array
          items:
            $ref: "#/components/schemas/ContentBlock"

    # ── CATEGORY ──────────────────────────────────────────────────────────────
...
    Enrollment:
      type: object
      required: [id, userId, courseId, status, progressPct, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        companyId:
          type: ["integer", "null"]
        employeeId:
          type: ["integer", "null"]
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        courseThumbnail:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, expired]
        progressPct:
          type: integer
        dueDate:
          type: ["string", "null"]
        assignmentStatus:
          $ref: "#/components/schemas/AssignmentStatus"
        lastAccessedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    EnrollmentDetail:
      allOf:
        - $ref: "#/components/schemas/Enrollment"
        - type: object
          properties:
            course:
              nullable: true
              allOf:
                - $ref: "#/components/schemas/CourseDetail"
    EnrollmentInput:
      type: object
      required: [courseId]
      properties:
        courseId:
          type: integer

    # ── PROGRESS ──────────────────────────────────────────────────────────────
...
    AssignmentStatus:
      type: string
      enum: [not_started, in_progress, completed, overdue]
...
    BlogPost:
      type: object
      required: [id, title, slug, excerpt, content, authorName, publishedAt, tags]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        authorName:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        publishedAt:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── TESTIMONIAL ───────────────────────────────────────────────────────────
...
    PlatformAdminUpdateLearningPathCoursesInput:
      type: object
      required: [courses]
      properties:
        courses:
          type: array
          items:
            type: object
            required: [courseId, position]
            properties:
              courseId:
                type: integer
              position:
                type: integer
              isRequired:
                type: boolean
...
    ContentBlock:
      type: object
      required: [id, type, position]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [heading, short_text, key_message, workplace_example, mauritian_example, practical_action, image, expandable, multiple_choice, decision_scenario, reflection, commitment]
        position:
          type: integer
        accessibilityLabel:
          type: string
        headingText:
          type: string
        bodyText:
          type: string
        imageUrl:
          type: string
        imageAlt:
          type: string
        expandableTitle:
          type: string
        expandableContent:
          type: string
        mcqQuestion:
          type: string
        mcqOptions:
          type: array
          items:
            type: string
        mcqCorrectIndex:
          type: integer
        mcqCorrectExplanation:
          type: string
        mcqIncorrectExplanation:
          type: string
        decisionIntro:
          type: string
        decisionPrompt:
          type: string
        decisionChoices:
          type: array
          items:
            $ref: "#/components/schemas/DecisionChoice"
        commitmentInstruction:
          type: string
        commitmentOptions:
          type: array
          items:
            $ref: "#/components/schemas/CommitmentOption"
    DecisionChoice:
      type: object
      required: [label, correct, feedback]
      properties:
        label:
          type: string
        correct:
          type: boolean
        feedback:
          type: string
    CommitmentOption:
      type: object
      required: [value, label, description]
      properties:
        value:
          type: string
        label:
          type: string
        description:
          type: string
...